[server] move usage -> org, stripe -> billing
