+++ components/server/src/container-module.ts
@@ -96,10 +96,10 @@ import { ContentServiceStorageClient } from "./storage/content-service-client";
 import { StorageClient } from "./storage/storage-client";
 import { AuthorizationService, AuthorizationServiceImpl } from "./user/authorization-service";
 import { NewsletterSubscriptionController } from "./user/newsletter-subscription-controller";
-import { StripeService } from "./user/stripe-service";
+import { StripeService } from "./billing/stripe-service";
 import { TokenProvider } from "./user/token-provider";
 import { TokenService } from "./user/token-service";
-import { UsageService } from "./user/usage-service";
+import { UsageService } from "./orgs/usage-service";
 import { ServerFactory, UserController } from "./user/user-controller";
 import { UserDeletionService } from "./user/user-deletion-service";
 import { UserService } from "./user/user-service";
